+++ internal/actions/tag.go
@@ -5,8 +5,8 @@ import (
 	"github.com/speakeasy-api/sdk-generation-action/internal/configuration"
 	"github.com/speakeasy-api/sdk-generation-action/internal/environment"
 	"github.com/speakeasy-api/sdk-generation-action/internal/logging"
+	"github.com/speakeasy-api/sdk-generation-action/internal/registry"
 	"golang.org/x/exp/maps"
-	"strings"
 )
 
 func Tag() error {
@@ -19,7 +19,7 @@ func Tag() error {
 		return err
 	}
 
-	tags := ProcessRegistryTags()
+	tags := registry.ProcessRegistryTags()
 
 	sources := environment.SpecifiedSources()
 	targets := environment.SpecifiedCodeSamplesTargets()
@@ -37,28 +37,4 @@ func Tag() error {
 	}
 
 	return cli.Tag(tags, sources, targets)
-}
-
-func ProcessRegistryTags() []string {
-	var tags []string
-	tagsInput := environment.RegistryTags()
-	if len(strings.Replace(tagsInput, " ", "", -1)) == 0 {
-		return tags
-	}
-
-	var processedTags []string
-	if strings.Contains(tagsInput, "\n") {
-		processedTags = strings.Split(tagsInput, "\n")
-	} else {
-		processedTags = strings.Split(tagsInput, ",")
-	}
-
-	for _, tag := range processedTags {
-		tag = strings.Replace(tag, " ", "", -1)
-		if len(tag) > 0 {
-			tags = append(tags, tag)
-		}
-	}
-
-	return tags
 }
